Address compiler warnings

+++ JoyShockLibrary/InputHelpers.cpp
@@ -67,13 +67,13 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 			}
 
 			// convert to real units
-			jc->imu_state.gyroX = (float)(gyroSampleX) * (2000.0 / 32767.0);
-			jc->imu_state.gyroY = (float)(gyroSampleY) * (2000.0 / 32767.0);
-			jc->imu_state.gyroZ = (float)(gyroSampleZ) * (2000.0 / 32767.0);
+			jc->imu_state.gyroX = (float)(gyroSampleX) * (2000.0f / 32767.0f);
+			jc->imu_state.gyroY = (float)(gyroSampleY) * (2000.0f / 32767.0f);
+			jc->imu_state.gyroZ = (float)(gyroSampleZ) * (2000.0f / 32767.0f);
 
-			jc->imu_state.accelX = (float)(accelSampleX) / 8192.0;
-			jc->imu_state.accelY = (float)(accelSampleY) / 8192.0;
-			jc->imu_state.accelZ = (float)(accelSampleZ) / 8192.0;
+			jc->imu_state.accelX = (float)(accelSampleX) / 8192.0f;
+			jc->imu_state.accelY = (float)(accelSampleY) / 8192.0f;
+			jc->imu_state.accelZ = (float)(accelSampleZ) / 8192.0f;
 
 			//printf("DS4 accel: %.4f, %.4f, %.4f\n", jc->imu_state.accelX, jc->imu_state.accelY, jc->imu_state.accelZ);
 
@@ -134,10 +134,10 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 			uint16_t stick2_y = packet[indexOffset+4];
 			stick2_y = 255 - stick2_y;
 
-			jc->simple_state.stickLX = (std::fmin)(1.0, (stick_x - 127.0) / 127.0);
-			jc->simple_state.stickLY = (std::fmin)(1.0, (stick_y - 127.0) / 127.0);
-			jc->simple_state.stickRX = (std::fmin)(1.0, (stick2_x - 127.0) / 127.0);
-			jc->simple_state.stickRY = (std::fmin)(1.0, (stick2_y - 127.0) / 127.0);
+			jc->simple_state.stickLX = (std::fmin)(1.0f, (stick_x - 127.0f) / 127.0f);
+			jc->simple_state.stickLY = (std::fmin)(1.0f, (stick_y - 127.0f) / 127.0f);
+			jc->simple_state.stickRX = (std::fmin)(1.0f, (stick2_x - 127.0f) / 127.0f);
+			jc->simple_state.stickRY = (std::fmin)(1.0f, (stick2_y - 127.0f) / 127.0f);
 
 			jc->push_sensor_samples(jc->imu_state.gyroX, jc->imu_state.gyroY, jc->imu_state.gyroZ,
 					jc->imu_state.accelX, jc->imu_state.accelY, jc->imu_state.accelZ, jc->delta_time);
@@ -179,13 +179,13 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
         }
 
         // convert to real units
-        jc->imu_state.gyroX = (float) (gyroSampleX) * (2000.0 / 32767.0);
-        jc->imu_state.gyroY = (float) (gyroSampleY) * (2000.0 / 32767.0);
-        jc->imu_state.gyroZ = (float) (gyroSampleZ) * (2000.0 / 32767.0);
+        jc->imu_state.gyroX = (float) (gyroSampleX) * (2000.0f / 32767.0f);
+        jc->imu_state.gyroY = (float) (gyroSampleY) * (2000.0f / 32767.0f);
+        jc->imu_state.gyroZ = (float) (gyroSampleZ) * (2000.0f / 32767.0f);
 
-        jc->imu_state.accelX = (float) (accelSampleX) / 8192.0;
-        jc->imu_state.accelY = (float) (accelSampleY) / 8192.0;
-        jc->imu_state.accelZ = (float) (accelSampleZ) / 8192.0;
+        jc->imu_state.accelX = (float) (accelSampleX) / 8192.0f;
+        jc->imu_state.accelY = (float) (accelSampleY) / 8192.0f;
+        jc->imu_state.accelZ = (float) (accelSampleZ) / 8192.0f;
 
         //printf("DS accel: %.4f, %.4f, %.4f\n", jc->imu_state.accelX, jc->imu_state.accelY, jc->imu_state.accelZ);
 
@@ -250,10 +250,10 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 		uint16_t stick2_y = packet[indexOffset + 3];
 		stick2_y = 255 - stick2_y;
 
-		jc->simple_state.stickLX = (std::fmin)(1.0, (stick_x - 127.0) / 127.0);
-		jc->simple_state.stickLY = (std::fmin)(1.0, (stick_y - 127.0) / 127.0);
-		jc->simple_state.stickRX = (std::fmin)(1.0, (stick2_x - 127.0) / 127.0);
-		jc->simple_state.stickRY = (std::fmin)(1.0, (stick2_y - 127.0) / 127.0);
+		jc->simple_state.stickLX = (std::fmin)(1.0f, (stick_x - 127.0f) / 127.0f);
+		jc->simple_state.stickLY = (std::fmin)(1.0f, (stick_y - 127.0f) / 127.0f);
+		jc->simple_state.stickRX = (std::fmin)(1.0f, (stick2_x - 127.0f) / 127.0f);
+		jc->simple_state.stickRY = (std::fmin)(1.0f, (stick2_y - 127.0f) / 127.0f);
 
 		jc->push_sensor_samples(jc->imu_state.gyroX, jc->imu_state.gyroY, jc->imu_state.gyroZ,
 				jc->imu_state.accelX, jc->imu_state.accelY, jc->imu_state.accelZ, jc->delta_time);
@@ -469,7 +469,7 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 			jc->simple_state.buttons |= ((buttons_pressed >> 8) << JSOFFSET_MINUS) & JSMASK_MINUS;
 			jc->simple_state.buttons |= ((buttons_pressed >> 6) << JSOFFSET_L) & JSMASK_L;
 			jc->simple_state.buttons |= ((buttons_pressed >> 13) << JSOFFSET_CAPTURE) & JSMASK_CAPTURE;
-			jc->simple_state.lTrigger = (buttons_pressed >> 7) & 1;
+			jc->simple_state.lTrigger = (float)((buttons_pressed >> 7) & 1);
 			jc->simple_state.buttons |= ((int)(jc->simple_state.lTrigger) << JSOFFSET_ZL) & JSMASK_ZL;
 			jc->simple_state.buttons |= ((buttons_pressed >> 5) << JSOFFSET_SL) & JSMASK_SL;
 			jc->simple_state.buttons |= ((buttons_pressed >> 4) << JSOFFSET_SR) & JSMASK_SR;
@@ -488,7 +488,7 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 			jc->simple_state.buttons |= ((buttons_pressed >> 25) << JSOFFSET_PLUS) & JSMASK_PLUS;
 			jc->simple_state.buttons |= ((buttons_pressed >> 22) << JSOFFSET_R) & JSMASK_R;
 			jc->simple_state.buttons |= ((buttons_pressed >> 28) << JSOFFSET_HOME) & JSMASK_HOME;
-			jc->simple_state.rTrigger = (buttons_pressed >> 23) & 1;
+			jc->simple_state.rTrigger = (float)((buttons_pressed >> 23) & 1);
 			jc->simple_state.buttons |= ((int)(jc->simple_state.rTrigger) << JSOFFSET_ZR) & JSMASK_ZR;
 			jc->simple_state.buttons |= ((buttons_pressed >> 21) << JSOFFSET_SL) & JSMASK_SL;
 			jc->simple_state.buttons |= ((buttons_pressed >> 20) << JSOFFSET_SR) & JSMASK_SR;
@@ -521,8 +521,8 @@ bool handle_input(JoyShock *jc, uint8_t *packet, int len, bool &hasIMU) {
 			jc->simple_state.buttons |= ((buttons_pressed >> 6) << JSOFFSET_L) & JSMASK_L;
 			jc->simple_state.buttons |= ((buttons_pressed >> 28) << JSOFFSET_HOME) & JSMASK_HOME;
 			jc->simple_state.buttons |= ((buttons_pressed >> 13) << JSOFFSET_CAPTURE) & JSMASK_CAPTURE;
-			jc->simple_state.rTrigger = (buttons_pressed >> 23) & 1;
-			jc->simple_state.lTrigger = (buttons_pressed >> 7) & 1;
+			jc->simple_state.rTrigger = (float)((buttons_pressed >> 23) & 1);
+			jc->simple_state.lTrigger = (float)((buttons_pressed >> 7) & 1);
 			jc->simple_state.buttons |= ((int)(jc->simple_state.lTrigger) << JSOFFSET_ZL) & JSMASK_ZL;
 			jc->simple_state.buttons |= ((int)(jc->simple_state.rTrigger) << JSOFFSET_ZR) & JSMASK_ZR;
